+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# v1.8.2 (not yet released)
+
+### `@liveblocks/react-comments`
+
+- Improve network loss and document visibility handling during comment
+  revalidation.
+- Better error handling on mutations (e.g. thread creation, comment creation,
+  etc.)
+
 # v1.8.1
 
 - Fix a bug in `toPlainLson` helper